Aborting changes
Virtualizing the table view is promissing, however the changes are not
so trivial and will need more time.

Changing the line array in Table from having all the lines in view to
having only the displayed lines has many side effect especially inside
the Cursor struct that uses that for its selections.

Will keep this branch as it can be used in the future if we want to
retry this method, we'll just need more time.