[editor] Improved partial prefab removal
Added back the (editor)removeInstance function.
This function was removed when switching to prefab2, removing the
ability for the editor to remove parts of the prefab tree without
rebuilding everything.

This in turn caused problems in the Terrain editor because changes to
the terrain were stored in the local3d object, which was destroyed
if the scene were to be refreshed, like by removing another object
from the scene, causing loss of work.