Hide kit (#280)
This introduce the hide.kit framework, that should now be used to create the inspectors interfaces for prefab in the editor. This is the first step needed to allow a smooth porting of the hide editor to the heaps target instead of js. This framework will allow interfaces to both work in the js version and heaps version by removing the dependency on jquery and the js dom in general.